Add Docker Compose support with localstack service
This commit introduces Docker Compose support by adding a `docker-compose.yml` file. The file includes configuration for LocalStack 2.0, enabling local development and testing against AWS cloud services. A volume is also defined for LocalStack to provide persistent storage. This volume path is added to the `.gitignore` file to prevent it from being committed to the repository.

Additionally, the `pom.xml` is updated to include the `spring-boot-docker-compose` dependency, providing support for Docker Compose within the Spring Boot application. Note that an unused property was removed from this file as well.