Switch SQS client from async to sync and modify environment variable
This commit replaces the asynchronous SQS client with a synchronous one in the SqsService class. This was done to align the client type with standard AWS practices. The change necessitated refactoring the methods using the client, which have been updated accordingly in the SqsService class.

Furthermore, the environment variable in application.properties and Dockerfile was renamed from 'AWS_END_POINT' to 'SQS_ENDPOINT_URL' to better reflect its purpose.

For better debugging, log level for spring-cloud library was also added.

Lastly, all commented code in AwsConfig.java was removed to enhance code readability and maintainability.